pyproject.toml
[project]
name = "autooptettinx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets<4.0",
    "dspy-ai>=2.6.27",
    "rich>=14.1.0",
    "sentence-transformers>=5.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
